fixing lossy

+++ 68_knight/KnightTour.java
@@ -72,8 +72,8 @@ public class KnightTour
 
     //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     //for random starting location, use lines below:
-    int startX = (Math.random() * n) + 2;
-    int startY = (Math.random() * n) + 2;
+    int startX = (int) (Math.random() * n) + 2;
+    int startY = (int) (Math.random() * n) + 2;
     //tf.findTour( startX, startY, 1 );   // 1 or 0 ?
     //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
